Fix 'implicit conversion from float to double' errors. (part 3)
These warnings showed up when enabling the '-Wdouble-promotion' GCC
compiler flag.  This batch of fixes consists of explicitly converting
from float to double.  These changes were made to match what seemed to
be the intent of the calculations.

+++ mythtv/libs/libmythtv/avformatdecoder.cpp
@@ -125,7 +125,8 @@ static float get_aspect(const AVCodecContext &ctx)
 
     if (ctx.sample_aspect_ratio.num && ctx.height)
     {
-        aspect_ratio = av_q2d(ctx.sample_aspect_ratio) * (float) ctx.width;
+        aspect_ratio = av_q2d(ctx.sample_aspect_ratio) *
+            static_cast<double>(ctx.width);
         aspect_ratio /= (float) ctx.height;
     }
